requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
BigNumber==1
bitarray==1.2.2
bleach==5.0.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.11
cytoolz==0.11.2
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
executing==0.9.1
fastjsonschema==2.16.1
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
lru-dict==1.1.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
packaging==21.3
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.14.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.27.1
rlp==2.0.1
Send2Trash==1.8.0
six==1.16.0
snakeviz==2.1.1
soupsieve==2.3.2.post1
stack-data==0.3.0
terminado==0.15.0
tinycss2==1.1.1
toolz==0.11.2
tornado==6.1
traitlets==5.3.0
urllib3==1.26.8
varint==1.0.2
wcwidth==0.2.5
web3==5.27.0
webencodings==0.5.1
websockets==9.1
widgetsnbextension==3.6.1
yarl==1.7.2
python-dotenv==1.1.0
